[DebugInfo][LLDB] Add support for two byte addresses
This change allows llvm-objdump and LLDB to parse AVR binaries.

In AVR, addresses are normally just 2 bytes in size, and in fact AVR has
set CodePointerSize to 2. GCC, on the other hand, uses 4 bytes for this,
probably to support larger devices such as the atmega2560. For now I
think it would be best to keep the CodePointerSize at 2 and make sure
LLVM tools support that.