[AVR] Optimize call frame
This is not just an optimization, it also fixes a bug I previously
introduced while fixing a different bug
(https://reviews.llvm.org/rG5aa8014ca811c8132ff99921c1eb39940f98c7e9).
In short, while the register R31R30 would be clobbered by the call, it
could be assigned by the register allocator in the call frame setup
code. That would overwrite the address of the stack pointer (stored in
R31R30) which would of course lead to undefined behavior.

This patch replaces that code entirely, replacing the stack adjustment
and stack-relative store instructions with plain push instructions. It
makes sure to work correctly even with undef values (by pushing the zero
register), which do not result in stack pushes but still need to adjust
the stack.