[AVR] Replace shift-to-loop IR pass with common shift code
I wrote a late IR pass a while ago to replace non-constant shift
operations with a constant shift in a loop (see
https://reviews.llvm.org/D96677). This worked, but wasn't optimal and
there is a possibility later passes could again re-introduce such shift
operations. But as a stopgap, it worked well enough.

Now that constant 32-bit shift operations are optimized in a later
stage, we can use the same code to lower shift operations to loops.

This patch removes that IR pass and replaces it with one that runs
during ISel lowering, like the constant shift operations.

I did not expect this to affect binary size in any significant way, but
apparently it does. The compiler-rt library (when compiled inside
TinyGo) becomes 1.8% smaller, and picolibc becomes 0.4% smaller. That's
a nice win for what was intended to be mostly just a refactor.